requirements.txt
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
googletrans==4.0.0rc1
textblob==0.17.1
nltk==3.8.1
pandas==2.1.1
plotly==5.17.0
python-dotenv==1.0.0
